Return the whole ApiResponse instead of just the data.
While initially I was going to make some form of error handler for crypto.com errors, it felt biased to do so and I should allow the users to see and handle the errors themselves. I may still add the ability for those errors to be returned as a enum of possible errors though instead of just the raw number values of the errors.